requirements.txt
black==25.1.0
click==8.1.8
dnspython==2.7.0
mypy-extensions==1.0.0
numpy==2.2.4
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pymongo==4.11.2
python-dotenv==1.0.1
